ajout de filtre et btn visible ou non visible

+++ SideBar Nav/ViewModel/FactureViewModel.cs
@@ -1,36 +1,61 @@
-﻿namespace TheClassMain.ViewModel
+﻿using System;
+using System.Collections.ObjectModel;
+using System.ComponentModel;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Data;
+using Microsoft.EntityFrameworkCore;
+using TheClassMain.Model;
+using TheClassMain.Query;
+using TheClassMain.Service;
+
+namespace TheClassMain.ViewModel
 {
-    using System;
-    using System.Collections.ObjectModel;
-    using System.ComponentModel;
-    using System.Linq;
-    using System.Runtime.CompilerServices;
-    using System.Threading.Tasks;
-    using System.Windows;
-    using Microsoft.EntityFrameworkCore;
-    using TheClassMain.Model;
-    using TheClassMain.Query;
-    using TheClassMain.Service;
-    public class FactureViewModel : INotifyPropertyChanged
+    public class FactureViewModel : ViewModelBase
     {
-        public ObservableCollection<Factures> facturesList = new ObservableCollection<Factures>();
+        private ObservableCollection<Factures> facturesList = new ObservableCollection<Factures>();
+        private ObservableCollection<Categories> categoriesList = new ObservableCollection<Categories>();
+        private Factures selectedFacture;
+        private Categories selectedCategorie;
+        private string description;
+        private string montant;
+        private DateTime? date;
+        private int userFactureId;
+        private string _facturesFilter = string.Empty;
+        private Visibility _btnVisibility = Visibility.Hidden;
+        public ICollectionView FacturesCollectionView { get; }
+        
+        public string FacturesFilter
+        {
+            get => _facturesFilter;
+            set
+            {
+                _facturesFilter = value;
+                OnPropertyChanged(nameof(FacturesFilter));
+                FacturesCollectionView.Refresh();
+            }
+        }
 
-        public ObservableCollection<Categories> categoriesList = new ObservableCollection<Categories>();
+        public FactureViewModel()
+        {
+            FacturesCollectionView = CollectionViewSource.GetDefaultView(facturesList);
+            FacturesCollectionView.Filter = FilterFactures;
+        }
 
-        public Factures selectedFacture;
+        private bool FilterFactures(object obj)
+        {
+            if (obj is Factures facture)
+            {
+                return (facture.Description?.IndexOf(FacturesFilter, StringComparison.OrdinalIgnoreCase) >= 0)
+                    || (facture.CategorieName?.IndexOf(FacturesFilter, StringComparison.OrdinalIgnoreCase) >= 0);
+            }
+            return false;
+        }
 
-        public Categories selectedCategorie;
-
-        public string description;
-
-        public string montant;
-
-        public DateTime? date;
-
-        public int userFactureId;
         public ObservableCollection<Factures> FacturesList => facturesList;
-
         public ObservableCollection<Categories> CategoriesList => categoriesList;
+
         public Factures SelectedFacture
         {
             get => selectedFacture;
@@ -40,7 +65,13 @@
         public Categories SelectedCategorie
         {
             get => selectedCategorie;
-            set { selectedCategorie = value; OnPropertyChanged(); }
+            set { selectedCategorie = value; OnPropertyChanged();}
+        }
+
+        public Visibility BtnVisibility
+        {
+            get => _btnVisibility;
+            set { _btnVisibility = value; OnPropertyChanged(); }
         }
 
         public string Description
@@ -78,18 +109,17 @@
             foreach (var facture in factures)
                 facturesList.Add(facture);
         }
+
         public async Task LoadCategories()
         {
-            using (var context = new TableContext())
-            {
-                var cats = await context.CategoriesT
-                    .Where(c => c.CustomerId == Session.CurrentCustomer.CustomerId && c.IsActive)
-                    .ToListAsync();
+            using var context = new TableContext();
+            var cats = await context.CategoriesT
+                .Where(c => c.CustomerId == Session.CurrentCustomer.CustomerId && c.IsActive)
+                .ToListAsync();
 
-                categoriesList.Clear();
-                foreach (var cat in cats)
-                    categoriesList.Add(cat);
-            }
+            categoriesList.Clear();
+            foreach (var cat in cats)
+                categoriesList.Add(cat);
         }
 
         public async Task AddFacture()
@@ -122,35 +152,55 @@
         public async Task UpdateFacture()
         {
             if (SelectedFacture == null || !ValidateInputs(out decimal parsedMontant))
-                return;
-            using var context = new TableContext();
-            var factureToUpdate = await context.FacturesT.FindAsync(SelectedFacture.FactureId);
-            if (factureToUpdate != null)
             {
-                factureToUpdate.Description = Description;
-                factureToUpdate.Montant = parsedMontant;
-                factureToUpdate.Date = Date.Value;
-                factureToUpdate.CategorieId = SelectedCategorie.CategorieId;
-                factureToUpdate.CategorieName = SelectedCategorie.Name;
-                await context.SaveChangesAsync();
-                await LoadFactures();
-                ClearInputs();
+                MessageBox.Show($"Please select a facture first", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+            else
+            {
+                using var context = new TableContext();
+                var factureToUpdate = await context.FacturesT.FindAsync(SelectedFacture.FactureId);
+                if (factureToUpdate != null)
+                {
+                    factureToUpdate.Description = Description;
+                    factureToUpdate.Montant = parsedMontant;
+                    factureToUpdate.Date = Date.Value;
+                    factureToUpdate.CategorieId = SelectedCategorie.CategorieId;
+                    factureToUpdate.CategorieName = SelectedCategorie.Name;
+                    await context.SaveChangesAsync();
+                    await LoadFactures();
+                    ClearInputs();
+                }
             }
         }
+
         public async Task DeleteFacture()
         {
             if (SelectedFacture == null)
-                return;
-            using var context = new TableContext();
-            var factureToDelete = await context.FacturesT.FindAsync(SelectedFacture.FactureId);
-            if (factureToDelete != null)
             {
-                context.FacturesT.Remove(factureToDelete);
-                await context.SaveChangesAsync();
-                await LoadFactures();
-                ClearInputs();
+                MessageBox.Show($"Please select a facture first", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+            else
+            {
+                string message = "Are you sure you want to delete this facture?";
+                string title = "Delete";
+                using var context = new TableContext();
+                var factureToDelete = await context.FacturesT.FindAsync(SelectedFacture.FactureId);
+                if (factureToDelete != null)
+                {
+                    var res = MessageBox.Show(message, title, MessageBoxButton.YesNo, MessageBoxImage.Question);
+                    if (res == MessageBoxResult.Yes)
+                    {
+                        context.FacturesT.Remove(factureToDelete);
+                        await context.SaveChangesAsync();
+                        await LoadFactures();
+                        ClearInputs();
+                    }
+                }
             }
         }
+
         public bool ValidateInputs(out decimal parsedMontant)
         {
             parsedMontant = 0;
@@ -164,10 +214,12 @@
             }
             return true;
         }
+
         public void LoadSelectedFacture()
         {
             if (SelectedFacture != null)
             {
+                BtnVisibility = Visibility.Visible;
                 Description = SelectedFacture.Description;
                 Montant = SelectedFacture.Montant.ToString("0.00");
                 Date = SelectedFacture.Date;
@@ -182,10 +234,7 @@
             Date = null;
             SelectedCategorie = null;
             SelectedFacture = null;
+            BtnVisibility = Visibility.Hidden;
         }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-        protected void OnPropertyChanged([CallerMemberName] string name = null) =>
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
     }
 }
